+++ src/Exception/Util.php
@@ -32,6 +32,10 @@ final class Util
 
     public static function wrap(\InvalidArgumentException $e)
     {
+        if ($e instanceof InvalidArgumentException) {
+            return $e;
+        }
+
         return new InvalidArgumentException($e->getMessage(), $e->getCode(), $e);
     }
 }