+++ tests/Spec/DefaultSpecFactoryTest.php
@@ -20,7 +20,7 @@ final class DefaultSpecFactoryTest extends \PHPUnit_Framework_TestCase
      * @dataProvider dataCreateAtDate
      * @requires function error_clear_last
      */
-    public function testCreateAtDate($expectedSpec, \DateTime $now)
+    public function testCreateAtDateDeprecated($expectedSpec, \DateTime $now)
     {
         error_clear_last();
 
@@ -28,11 +28,28 @@ final class DefaultSpecFactoryTest extends \PHPUnit_Framework_TestCase
         $error = error_get_last();
 
         $this->assertArraySubset([
-            'message' => 'Cs278\\BankModulus\\Spec\\DefaultSpecFactory::withNow() is for testing only!',
-            'type' => E_USER_WARNING,
+            'message' => 'Cs278\\BankModulus\\Spec\\DefaultSpecFactory::withNow() is deprecated use withDate() instead. Note this method was marked @internal maybe removed in a minor release.',
+            'type' => E_USER_DEPRECATED,
         ], $error);
 
         error_clear_last();
+
+        // Run tests 5 times to ensure consistent results.
+        for ($i = 0; $i < 5; ++$i) {
+            $spec = $factory->create();
+
+            $this->assertInstanceOf('Cs278\\BankModulus\\Spec\\SpecInterface', $spec);
+            $this->assertInstanceOf('Cs278\\BankModulus\\Spec\\'.$expectedSpec, $spec);
+        }
+    }
+
+    /**
+     * @dataProvider dataCreateAtDate
+     */
+    public function testCreateAtDate($expectedSpec, \DateTime $now)
+    {
+        $factory = new DefaultSpecFactory();
+        $factory = $factory->withDate($now);
 
         // Run tests 5 times to ensure consistent results.
         for ($i = 0; $i < 5; ++$i) {
@@ -73,6 +90,85 @@ final class DefaultSpecFactoryTest extends \PHPUnit_Framework_TestCase
             ['VocaLinkV500', new \DateTime('2018-09-16')],
             ['VocaLinkV510', new \DateTime('2018-09-17')],
             ['VocaLinkV510', new \DateTime('2030-01-01')],
+        ];
+    }
+
+    public function testWithDateDateTime()
+    {
+        $factory = new DefaultSpecFactory();
+        $newFactory = $factory->withDate(new \DateTime('2018-01-01'));
+
+        $this->assertNotSame($factory, $newFactory);
+        $this->assertInstanceOf('Cs278\\BankModulus\\Spec\\VocaLinkV460', $newFactory->create());
+    }
+
+    /**
+     * @requires PHP 5.5.0
+     */
+    public function testWithDateDateTimeImmutable()
+    {
+        $factory = new DefaultSpecFactory();
+        $newFactory = $factory->withDate(new \DateTimeImmutable('2018-01-01'));
+
+        $this->assertNotSame($factory, $newFactory);
+        $this->assertInstanceOf('Cs278\\BankModulus\\Spec\\VocaLinkV460', $newFactory->create());
+    }
+
+    /**
+     * @expectedException \Cs278\BankModulus\Exception\InvalidArgumentException
+     * @expectedExceptionMessage Expected an instance of DateTimeInterface. Got: stdClass
+     * @requires PHP 5.5
+     */
+    public function testWithDateObjectInvalid()
+    {
+        $factory = new DefaultSpecFactory();
+        $factory->withDate(new \stdClass());
+    }
+
+    /**
+     * @expectedException \Cs278\BankModulus\Exception\InvalidArgumentException
+     * @expectedExceptionMessage Expected an instance of DateTime. Got: stdClass
+     */
+    public function testWithDateObjectInvalidPhp54()
+    {
+        if (PHP_VERSION_ID >= 50500) {
+            $this->markTestSkipped('Test requires PHP < 5.5');
+        }
+
+        $factory = new DefaultSpecFactory();
+        $factory->withDate(new \stdClass());
+    }
+
+    public function testWithDateString()
+    {
+        $factory = new DefaultSpecFactory();
+        $newFactory = $factory->withDate('2018-01-01');
+
+        $this->assertNotSame($factory, $newFactory);
+        $this->assertInstanceOf('Cs278\\BankModulus\\Spec\\VocaLinkV460', $newFactory->create());
+    }
+
+    /**
+     * @dataProvider dataWithDateStringInvalid
+     * @expectedException \Cs278\BankModulus\Exception\InvalidArgumentException
+     * @expectedExceptionMessage Expecting valid ISO8601 date (YYYY-MM-DD). Got:
+     */
+    public function testWithDateStringInvalid($input)
+    {
+        $factory = new DefaultSpecFactory();
+        $factory->withDate($input);
+    }
+
+    public function dataWithDateStringInvalid()
+    {
+        return [
+            [''],
+            [' 2018-01-01'],
+            ['2018-01-01 '],
+            [' 2018-01-01 '],
+            ['18-01-01'],
+            ['ab18-01-01'],
+            ['2018-02-30'],
         ];
     }
 }